Change output to html_document2.
Bookdown has an output format designed for
single pages called html_document2. These
pages are self-contained by default, so the
lines to avoid splitting or creating extra
files are unnecessary.

Hopefully, the switch will also help address
the other problem with the gitbook format:
namely, that the book won't output equations
if the document is self-contained.