Add protein purification calculations
Created an RMarkdown document that outlines all the
calculations that would need to be performed in the
protein purification experiment. The document is
written to use the bookdown package for knitting,
but it wouldn't be knittable as is, since there
is a file choice in the code, and only one MW has
been added to the data frame.