Adjust EnzymeKinetics.Rmd for units package
The EnzymeKinetics.Rmd file was rewritten to
incorporate extensive use of the 'units'
package. The lines that record values used
in making solutions were adjusted, lines
for setting units on the data frames were added,
and accomodation for the incompatibility of
the units package and curve fitting was made.
The file also takes advantage of the ability
to add units to the axis labels for graphs. The
file was tested with an example data set, not
included.